LayoutTests/platform/ios-simulator/fast/dom/Window/window-properties-device-orientation-expected.txt

CONSOLE MESSAGE: Deprecated attempt to access property 'bubbles' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'cancelBubble' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'cancelable' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'currentTarget' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'defaultPrevented' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'eventPhase' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'returnValue' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'srcElement' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'target' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'timeStamp' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'type' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'bubbles' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'cancelBubble' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'cancelable' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'currentTarget' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'defaultPrevented' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'eventPhase' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'returnValue' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'srcElement' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'target' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'timeStamp' on a non-Event object.
CONSOLE MESSAGE: Deprecated attempt to access property 'type' on a non-Event object.
This test dumps all of the properties that are reachable from the window.DeviceMotionEvent window.ondevicemotion, window.DeviceOrientationEvent and window.ondeviceorientation objects, along with their types. These properties are currently guarded by ENABLE_DEVICE_ORIENTATION.

window.DeviceMotionEvent [object DeviceMotionEventConstructor]
window.DeviceMotionEvent.prototype [object DeviceMotionEventPrototype]
window.DeviceMotionEvent.prototype.AT_TARGET [number]
window.DeviceMotionEvent.prototype.BLUR [number]
window.DeviceMotionEvent.prototype.BUBBLING_PHASE [number]
window.DeviceMotionEvent.prototype.CAPTURING_PHASE [number]
window.DeviceMotionEvent.prototype.CHANGE [number]
window.DeviceMotionEvent.prototype.CLICK [number]
window.DeviceMotionEvent.prototype.DBLCLICK [number]
window.DeviceMotionEvent.prototype.DRAGDROP [number]
window.DeviceMotionEvent.prototype.FOCUS [number]
window.DeviceMotionEvent.prototype.KEYDOWN [number]
window.DeviceMotionEvent.prototype.KEYPRESS [number]
window.DeviceMotionEvent.prototype.KEYUP [number]
window.DeviceMotionEvent.prototype.MOUSEDOWN [number]
window.DeviceMotionEvent.prototype.MOUSEDRAG [number]
window.DeviceMotionEvent.prototype.MOUSEMOVE [number]
window.DeviceMotionEvent.prototype.MOUSEOUT [number]
window.DeviceMotionEvent.prototype.MOUSEOVER [number]
window.DeviceMotionEvent.prototype.MOUSEUP [number]
window.DeviceMotionEvent.prototype.NONE [number]
window.DeviceMotionEvent.prototype.SELECT [number]
window.DeviceMotionEvent.prototype.bubbles [undefined]
window.DeviceMotionEvent.prototype.cancelBubble [undefined]
window.DeviceMotionEvent.prototype.cancelable [undefined]
window.DeviceMotionEvent.prototype.currentTarget [undefined]
window.DeviceMotionEvent.prototype.defaultPrevented [undefined]
window.DeviceMotionEvent.prototype.eventPhase [undefined]
window.DeviceMotionEvent.prototype.initDeviceMotionEvent [function]
window.DeviceMotionEvent.prototype.initEvent [function]
window.DeviceMotionEvent.prototype.preventDefault [function]
window.DeviceMotionEvent.prototype.returnValue [undefined]
window.DeviceMotionEvent.prototype.srcElement [undefined]
window.DeviceMotionEvent.prototype.stopImmediatePropagation [function]
window.DeviceMotionEvent.prototype.stopPropagation [function]
window.DeviceMotionEvent.prototype.target [undefined]
window.DeviceMotionEvent.prototype.timeStamp [undefined]
window.DeviceMotionEvent.prototype.type [undefined]
window.ondevicemotion [null]
window.DeviceOrientationEvent [object DeviceOrientationEventConstructor]
window.DeviceOrientationEvent.prototype [object DeviceOrientationEventPrototype]
window.DeviceOrientationEvent.prototype.AT_TARGET [number]
window.DeviceOrientationEvent.prototype.BLUR [number]
window.DeviceOrientationEvent.prototype.BUBBLING_PHASE [number]
window.DeviceOrientationEvent.prototype.CAPTURING_PHASE [number]
window.DeviceOrientationEvent.prototype.CHANGE [number]
window.DeviceOrientationEvent.prototype.CLICK [number]
window.DeviceOrientationEvent.prototype.DBLCLICK [number]
window.DeviceOrientationEvent.prototype.DRAGDROP [number]
window.DeviceOrientationEvent.prototype.FOCUS [number]
window.DeviceOrientationEvent.prototype.KEYDOWN [number]
window.DeviceOrientationEvent.prototype.KEYPRESS [number]
window.DeviceOrientationEvent.prototype.KEYUP [number]
window.DeviceOrientationEvent.prototype.MOUSEDOWN [number]
window.DeviceOrientationEvent.prototype.MOUSEDRAG [number]
window.DeviceOrientationEvent.prototype.MOUSEMOVE [number]
window.DeviceOrientationEvent.prototype.MOUSEOUT [number]
window.DeviceOrientationEvent.prototype.MOUSEOVER [number]
window.DeviceOrientationEvent.prototype.MOUSEUP [number]
window.DeviceOrientationEvent.prototype.NONE [number]
window.DeviceOrientationEvent.prototype.SELECT [number]
window.DeviceOrientationEvent.prototype.bubbles [undefined]
window.DeviceOrientationEvent.prototype.cancelBubble [undefined]
window.DeviceOrientationEvent.prototype.cancelable [undefined]
window.DeviceOrientationEvent.prototype.currentTarget [undefined]
window.DeviceOrientationEvent.prototype.defaultPrevented [undefined]
window.DeviceOrientationEvent.prototype.eventPhase [undefined]
window.DeviceOrientationEvent.prototype.initDeviceOrientationEvent [function]
window.DeviceOrientationEvent.prototype.initEvent [function]
window.DeviceOrientationEvent.prototype.preventDefault [function]
window.DeviceOrientationEvent.prototype.returnValue [undefined]
window.DeviceOrientationEvent.prototype.srcElement [undefined]
window.DeviceOrientationEvent.prototype.stopImmediatePropagation [function]
window.DeviceOrientationEvent.prototype.stopPropagation [function]
window.DeviceOrientationEvent.prototype.target [undefined]
window.DeviceOrientationEvent.prototype.timeStamp [undefined]
window.DeviceOrientationEvent.prototype.type [undefined]
window.ondeviceorientation [null]
